Complete checkpoint 1 & 2 - load artefacts
Loads the provided file of the museum's artefacts and displays the data, and let's you change the description of the items.
Extend the file of the museum's artefacts with a filename for each artefact referring to an image of the artefact. Then with your application display this image to the user.